requirements.txt
Package      Version
------------ ---------
certifi      2022.9.24
greenlet     2.0.1
pip          22.2.2
psycopg2     2.9.5
PyYAML       6.0
setuptools   65.5.0
SQLAlchemy   1.4.44
wheel        0.37.1
wincertstore 0.2
